requirements.txt
streamlit
tensorflow
pillow
numpy
pandas
huggingface_hub
scikit-learn
